refactor(data-model): share the codec benchmark subjects, and add two series
The subjects move out of `codec-json.bench.ts` into `codec-fixtures.ts`, so
that a second format's benchmark measures the same values rather than a set
written separately beside it. Only then is a difference between two formats'
numbers a difference between the formats. The move is faithful: every code
line is unchanged bar the new import and one rewrap, and the one paragraph
that made a JSON-specific claim about the sparse series now says which format
it is claiming it of.

Two omnibus series join the existing one, both defined by what a format has to
do with them rather than by their size:

* `json-pass-through` holds only types JSON carries directly, so it draws no
  tag and no escape under any format. What a walk costs over it is the walk.
* `realm-pass-through` adds the five things JSON must write out as tagged text
  or escape around -- `bigint`, `undefined`, the special numbers, a
  `/`-prefixed key, and an array hole -- and nothing else. A realm-crossing
  format carries the whole of it directly.

The gap between the two is the tagging and escaping machinery stated as a
measurement, and the second series is the territory a second format buys.
Verified rather than asserted: the first emits no tag at all, and the second
emits exactly `/BigInt@1`, `/SpecialNumber@1`, `/Undefined@1`, `/hole` and
`/object`.
